adeguard_streamlit_dashboard/requirements.txt
# ADEGuard Streamlit Dashboard Requirements
# Current Date: 2025-10-17 18:29:22
# User: ghanashyam9348

plotly==5.17.0
requests==2.31.0
python-dateutil==2.8.2
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4
altair==5.2.0